+++ src/main/java/de/tum/cit/aet/artemis/plagiarism/dto/PlagiarismCaseUserDTO.java
@@ -2,7 +2,7 @@ package de.tum.cit.aet.artemis.plagiarism.dto;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
 
-import de.tum.cit.aet.artemis.core.domain.User;
+import de.tum.cit.aet.artemis.account.domain.User;
 
 @JsonInclude(JsonInclude.Include.NON_EMPTY)
 public record PlagiarismCaseUserDTO(Long id, String login, String name, String visibleRegistrationNumber) {
